requirements.txt
streamlit==1.28.0
yt-dlp==2023.10.13
moviepy==1.0.3
python-dotenv==1.0.0
google-api-python-client==2.100.0
youtube-transcript-api==0.6.1
Pillow>=11.0.0
numpy==2.3.1
librosa>=0.10.2
numpy==2.1.3
# your other packages...
streamlit==1.29.0
moviepy==1.0.3
opencv-python==4.8.0.74
ffmpeg-python==0.2.1
openai==1.30.0
faster-whisper==0.5.1
elevenlabs==0.2.3
numpy==1.26.0
